Server-side hot reloading
The important part of this change is in startDev.js the loading of
routes using a getter, every time we need routes is that main change.
The other big change is accepting every get request and processing it
through processRoute. If we specify every URL then we can't cache bust
the server side. This however has a trade-off - 404s aren't handled
properly. They need to be special cased in processRoute (next commit).
