+++ src/cli/commands/startDev.js
@@ -1,34 +1,34 @@
 /* eslint-disable global-require */
 const path = require('path');
 const dotenv = require('dotenv');
-// const chokidar = require('chokidar');
+const chokidar = require('chokidar');
 require('isomorphic-fetch');
 require('babel-polyfill');
 require('babel-register')(require('../../config/babelrcHelper').default(true, process.cwd(), true));
 const startApp = require('../../server/server').default;
 
 module.exports = (args) => {
-  let routes;
-  // let watcher;
+  let watcher;
 
   if (args._[1]) {
     // eslint-disable-next-line global-require, import/no-dynamic-require
-    routes = require(`${process.cwd()}/${args._[1]}/src/urls`).default;
-    // watcher = chokidar.watch(`${process.cwd()}/${args._[1]}`);
+    watcher = chokidar.watch(`${process.cwd()}/${args._[1]}`);
     dotenv.config({ path: path.join(process.cwd(), args._[1], '.env') });
   } else {
     // eslint-disable-next-line global-require, import/no-dynamic-require
-    routes = require(`${process.cwd()}/src/urls`).default;
-    // watcher = chokidar.watch(`${process.cwd()}`);
+    watcher = chokidar.watch(`${process.cwd()}`);
     dotenv.config({ path: path.join(process.cwd(), '.env') });
   }
 
-  // TODO: this needs work
-  // watcher.on('ready', () => {
-  //   watcher.on('all', (event, filePath) => {
-  //     delete require.cache[filePath];
-  //   });
-  // });
+  watcher.on('ready', () => {
+    watcher.on('all', () => {
+      Object.keys(require.cache).forEach((id) => {
+        if (id.indexOf(`${process.cwd()}/src`) > -1) {
+          delete require.cache[id];
+        }
+      });
+    });
+  });
 
   return startApp({
     appName: process.env.APP_NAME,
@@ -38,6 +38,14 @@ module.exports = (args) => {
     host: process.env.HOST || 'http://localhost:3000',
     port: process.env.PORT || 3000,
     appLocation: process.cwd(),
-    routes,
+    get routes() {
+      if (args._[1]) {
+        // eslint-disable-next-line global-require, import/no-dynamic-require
+        return require(`${process.cwd()}/${args._[1]}/src/urls`).default;
+      }
+
+      // eslint-disable-next-line global-require, import/no-dynamic-require
+      return require(`${process.cwd()}/src/urls`).default;
+    },
   });
 };
